Refactor user authentication servlets and add logging configuration; update mock data handling and enhance feature documentation
